requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
blinker==1.8.2
build==1.2.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.14
executing==2.0.1
fastapi==0.112.2
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
iniconfig==2.0.0
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==30.1.0
langchain==0.2.15
langchain-chroma==0.1.3
langchain-community==0.2.15
langchain-core==0.2.37
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.2
langsmith==0.1.108
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.6.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
pluggy==1.5.0
posthog==3.6.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
replicate==0.32.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.38.2
streamlit==1.38.0
sympy==1.13.2
tenacity==8.5.0
tokenizers==0.20.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uv==0.4.1
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0.1
wrapt==1.16.0
yarl==1.9.6
zipp==3.20.1
